// frontend/shared/src/api/model/unitsType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Navtrack.Api
 * OpenAPI spec version: 1.0.0
 */

export type UnitsType = (typeof UnitsType)[keyof typeof UnitsType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UnitsType = {
  Metric: "Metric",
  Imperial: "Imperial"
} as const;